+++ appPlugins/ToolExtract.py
@@ -964,7 +964,7 @@ class ExtractUI:
         self.gerber_object_combo = FCComboBox()
         self.gerber_object_combo.setModel(self.app.collection)
         self.gerber_object_combo.setRootModelIndex(self.app.collection.index(0, 0, QtCore.QModelIndex()))
-        self.gerber_object_combo.is_last = True
+        self.gerber_object_combo.is_last = False
         self.gerber_object_combo.obj_type = "Gerber"
 
         self.tools_box.addWidget(self.gerber_object_combo)