+++ CHANGELOG.md
@@ -12,6 +12,7 @@ CHANGELOG for FlatCAM beta
 - fixed Extract Plugin not extracting soldermask for aperture macros
 - fixed INFO in Gerber Object UI to show correctly the type of geometric element in Follow geometry
 - in INFO functionality for Gerber Object made the difference if the Follow geometry is not a Point or a LineString
+- In Plugins: Panelize, Cutout, Extract, Copper Thieving disabled the autoload of the last created APP object
 
 18.09.2021